Add mobile menu for authentication and theme options
